Move SQLFunctionsIT as SQLIntegTestCase
