+++ src/test/java/com/amazon/opendistroforelasticsearch/sql/intgtest/MainTestSuite.java
@@ -72,12 +72,10 @@ import static org.mockito.Mockito.when;
     HashJoinBasicTest.class,
     HashJoinMoreTest.class,
     ExplainTest.class,
-    WktToGeoJsonConverterTests.class,
     SqlParserTests.class,
     ShowTest.class,
     CSVResultsExtractorTests.class,
     SourceFieldTest.class,
-    SQLFunctionsTest.class,
     UtilTests.class,
     MultiQueryTests.class,
     DeleteTest.class,
